move chunk management in the S3 part and remove useless chunk
