+++ util/syncador/Cargo.toml
@@ -36,6 +36,8 @@ tracing = { workspace = true }
 [dev-dependencies]
 uuid = { workspace = true }
 tempfile = { workspace = true }
+dot-movement = { workspace = true }
+tracing-subscriber = { workspace = true }
 
 [lints]
 workspace = true
