+++ util/syncador/src/backend/s3/shared_bucket/mod.rs
@@ -3,6 +3,8 @@ use aws_types::region::Region;
 use tracing::info;
 
 const UPLOAD_COMPLETE_MARKER_FILE_NAME: &str = "upload_complete.txt";
+pub(crate) const DEFAULT_CHUNK_SIZE: usize = 500 * 1024 * 1024; // 500 MB per chunk (adjustable)
+pub(crate) const BUFFER_SIZE: usize = 10 * 1024 * 1024; // 10 MB buffer for each read/write operation
 
 pub mod metadata;
 pub mod pull;
@@ -106,8 +108,122 @@ pub mod test {
 	//! pub in case we want to reuse helpers
 
 	use super::*;
+	use crate::backend::s3::bucket_connection::BucketConnection;
+	use crate::backend::s3::shared_bucket::pull::Pull;
+	use crate::backend::s3::shared_bucket::push::Push;
 	use crate::backend::{PullOperations, PushOperations};
 	use crate::files::package::{Package, PackageElement};
+	use movement_types::actor;
+	use std::fs::File;
+	use std::io::BufWriter;
+	use std::io::Write;
+
+	#[tokio::test]
+	pub async fn test_archive_split() -> Result<(), anyhow::Error> {
+		use tracing_subscriber::EnvFilter;
+
+		tracing_subscriber::fmt()
+			.with_env_filter(
+				EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info")),
+			)
+			.init();
+		// 1) Chunk size is bigger than the archive. No split in chunk.
+		process_archive_test("test_archive_split.tmp", 10 * 1024, 1024).await?;
+		// 2) Chunk size is smaller than the archive. Several chunk is create and reconstructed.
+		process_archive_test("test_archive_split2.tmp", 1024, 512).await?;
+		Ok(())
+	}
+
+	async fn process_archive_test(
+		temp_file_name: &str,
+		chunk_size: usize,
+		buffer_size: usize,
+	) -> Result<(), anyhow::Error> {
+		//Create source and destination temp dir.
+		let source_dir = tempfile::tempdir()?;
+		let destination_dir = tempfile::tempdir()?;
+
+		//create file to Push size 10 * 1024.
+		let archive_file_path = source_dir.path().join(temp_file_name);
+		{
+			let file = File::create(&archive_file_path)?;
+			let mut writer = BufWriter::new(file);
+			//Fill with some data. 10 Mb
+			let data: Vec<u8> = vec![2; 1024];
+			(0..10).try_for_each(|_| writer.write_all(&data))?;
+		}
+
+		let bucket = format!("public-test-bucket-{}", uuid::Uuid::new_v4());
+		let config = aws_config::load_from_env().await;
+		let client = aws_sdk_s3::Client::new(&config);
+
+		let connection = BucketConnection::new(client.clone(), bucket.clone());
+
+		connection.create_bucket_if_not_exists().await?;
+
+		let application_id = application::Id::new([
+			26, 43, 60, 77, 94, 111, 122, 139, 156, 173, 190, 207, 208, 225, 242, 3, 20, 37, 54,
+			71, 88, 105, 122, 139, 156, 173, 190, 207, 208, 225, 242, 3,
+		]);
+		let syncer_id = actor::Id::new([
+			10, 64, 193, 217, 99, 233, 100, 32, 31, 1, 244, 166, 56, 79, 213, 208, 112, 158, 162,
+			27, 10, 111, 130, 99, 91, 130, 103, 26, 12, 121, 210, 75,
+		]);
+
+		let metadata = metadata::Metadata::default()
+			.with_application_id(application_id)
+			.with_syncer_id(syncer_id);
+
+		let push = Push {
+			bucket_connection: connection,
+			metadata: metadata.clone(),
+			chunk_size,
+			buffer_size,
+		};
+
+		let element = PackageElement {
+			sync_files: vec![archive_file_path],
+			root_dir: source_dir.path().to_path_buf(),
+		};
+		let package = Package(vec![element]);
+		let archive_package = push.push(package).await?;
+
+		// Pull archive
+		let connection = BucketConnection::new(client.clone(), bucket.clone());
+		let pull = Pull::new(connection, metadata, destination_dir.path().to_path_buf());
+		let element = PackageElement {
+			sync_files: vec![archive_package.0[0].sync_files[0].clone()],
+			root_dir: destination_dir.path().to_path_buf(),
+		};
+		let package = Package(vec![element]);
+
+		let dest_package = pull
+			.pull(Some(package))
+			.await?
+			.ok_or(anyhow::anyhow!("Error No file pulled."))?;
+
+		let file_metadata = std::fs::metadata(&dest_package.0[0].sync_files[0])?;
+		let file_size = file_metadata.len() as usize;
+		assert_eq!(file_size, 10 * 1024, "dest file hasn't the right size: {file_size}");
+
+		//verify that all chunk has been removed
+		let has_chunk = std::fs::read_dir(&destination_dir)?
+			.find(|entry| {
+				entry
+					.as_ref()
+					.ok()
+					.map(|entry| {
+						let path = entry.path();
+						path.is_file()
+							&& path.extension().and_then(|ext| ext.to_str()) == Some("chunk")
+					})
+					.unwrap_or(false)
+			})
+			.is_some();
+		assert!(!has_chunk, "Some chunk are still present.");
+
+		Ok(())
+	}
 
 	#[tokio::test]
 	async fn test_create() -> Result<(), anyhow::Error> {
