+++ util/syncador/src/backend/s3/shared_bucket/pull.rs
@@ -1,11 +1,15 @@
 use super::metadata::Metadata;
 use crate::backend::s3::bucket_connection::BucketConnection;
+use crate::backend::s3::shared_bucket::BUFFER_SIZE;
 use crate::backend::PullOperations;
 use crate::files::package::{Package, PackageElement};
 use std::collections::{HashMap, HashSet};
+use std::fs::File;
+use std::fs::OpenOptions;
+use std::io::BufReader;
+use std::io::{Read, Write};
 use std::path::PathBuf;
 use tokio::io::AsyncWriteExt;
-use tracing::info;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Candidate {
@@ -108,37 +112,11 @@ impl Pull {
 		Ok(candidates.keys().cloned().collect())
 	}
 
-	/*pub(crate) async fn download_based_on_manifest(
-		&self,
-		candidate_selected: &Candidate,
-		manifest: PackageElement,
-	) -> Result<PackageElement, anyhow::Error> {
-		// get the path tuples
-		let path_tuples = manifest.try_path_tuples()?;
-
-		// download each file
-		let mut manifest_futures = Vec::new();
-		for (relative_path, full_path) in path_tuples {
-			let future = self.download_path(
-				candidate_selected,
-				relative_path.to_path_buf().clone(),
-				full_path.clone(),
-			);
-			manifest_futures.push(future);
-		}
-
-		// try to join all the manifest_futures
-		futures::future::try_join_all(manifest_futures).await?;
-
-		// should downloaded into the locations specified in the manifest
-		Ok(manifest)
-	}*/
-
 	pub(crate) async fn download_all_files_for_candidate(
 		&self,
 		candidate: &Candidate,
 	) -> Result<PackageElement, anyhow::Error> {
-		info!("Downloading all files for candidate: {:?}", candidate);
+		tracing::debug!("Downloading all files for candidate: {:?}", candidate);
 
 		// get all of the public file paths for this application
 		let public_file_paths = self
@@ -176,8 +154,22 @@ impl Pull {
 		// try to join all the manifest_futures
 		futures::future::try_join_all(manifest_futures).await?;
 
+		//recreate splited archive if needed.
+		let mut unsplit_manifest =
+			PackageElement { sync_files: vec![], root_dir: manifest.root_dir.clone() };
+		for absolute_path in &manifest.sync_files {
+			let path_buf = absolute_path.to_path_buf();
+			let absolute_path =
+				tokio::task::spawn_blocking(move || recreate_archive(path_buf)).await??;
+
+			if !unsplit_manifest.sync_files.contains(&absolute_path) {
+				tracing::info!("Archive file added {absolute_path:?}",);
+				unsplit_manifest.sync_files.push(absolute_path)
+			}
+		}
+
 		// should downloaded into the locations specified in the manifest
-		Ok(manifest)
+		Ok(unsplit_manifest)
 	}
 
 	async fn find_candidates(&self, package: &Package) -> Result<Vec<Candidate>, anyhow::Error> {
@@ -217,7 +209,7 @@ impl Pull {
 #[async_trait::async_trait]
 impl PullOperations for Pull {
 	async fn pull(&self, package: Option<Package>) -> Result<Option<Package>, anyhow::Error> {
-		info!("S3 pulling package: {:?}", package);
+		tracing::debug!("S3 pulling package: {:?}", package);
 		if package.is_none() {
 			return Ok(None);
 		}
@@ -233,4 +225,62 @@ impl PullOperations for Pull {
 		let candidate_selected = self.select_candidate_from(&package, candidates).await?;
 		Ok(Some(self.pull_candidate(package, candidate_selected).await?))
 	}
+}
+
+fn recreate_archive(archive_chunk: PathBuf) -> Result<PathBuf, anyhow::Error> {
+	if archive_chunk.extension().map(|ext| ext != "chunk").unwrap_or(true) {
+		//not a chunk file return.
+		return Ok(archive_chunk);
+	}
+
+	let archive_file_name = archive_chunk
+		.file_name()
+		.and_then(|file_name| file_name.to_str())
+		.and_then(|file_name_str| file_name_str.strip_suffix(".chunk"))
+		.and_then(|base_filename| {
+			let base_filename_parts: Vec<&str> = base_filename.rsplitn(2, '_').collect();
+			(base_filename_parts.len() > 1).then(|| base_filename_parts[1].to_string())
+		})
+		.ok_or(anyhow::anyhow!(format!(
+			"Archive filename not found for chunk path:{:?}",
+			archive_chunk.to_str()
+		)))?;
+
+	let archive_path = archive_chunk.parent().map(|parent| parent.join(archive_file_name)).ok_or(
+		anyhow::anyhow!(format!(
+			"Archive filename no root dir in path:{:?}",
+			archive_chunk.to_str()
+		)),
+	)?;
+
+	let mut archive_file = OpenOptions::new()
+		.create(true) // Create the file if it doesn't exist
+		.append(true) // Open in append mode (do not overwrite)
+		.open(&archive_path)?;
+
+	let mut buffer = vec![0; BUFFER_SIZE];
+
+	let chunk_file = File::open(&archive_chunk)?;
+	let mut chunk_reader = BufReader::new(chunk_file);
+
+	loop {
+		// Read a part of the chunk into the buffer
+		let bytes_read = chunk_reader.read(&mut buffer)?;
+
+		if bytes_read == 0 {
+			break; // End of chunk file
+		}
+
+		// Write the buffer data to the output file
+		archive_file.write_all(&buffer[..bytes_read])?;
+	}
+
+	let file_metadata = std::fs::metadata(&archive_path)?;
+	let file_size = file_metadata.len() as usize;
+
+	//remove the chunk that is useless.
+	std::fs::remove_file(&archive_chunk)?;
+	tracing::debug!("PULL {archive_path:?} archive_chunk size: {file_size}",);
+
+	Ok(archive_path)
 }
