+++ util/syncador/src/backend/s3/shared_bucket/push.rs
@@ -1,9 +1,14 @@
 use super::metadata::Metadata;
 use crate::backend::s3::bucket_connection::BucketConnection;
+use crate::backend::s3::shared_bucket::BUFFER_SIZE;
+use crate::backend::s3::shared_bucket::DEFAULT_CHUNK_SIZE;
 use crate::backend::PushOperations;
 use crate::files::package::{Package, PackageElement};
 use aws_sdk_s3::operation::put_object::PutObjectOutput;
 use aws_sdk_s3::primitives::ByteStream;
+use std::fs::File;
+use std::io::{BufReader as StdBufReader, Read, Write};
+use std::path::Path;
 use std::path::PathBuf;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
@@ -16,11 +21,18 @@ pub struct Candidate {
 pub struct Push {
 	pub bucket_connection: BucketConnection,
 	pub metadata: Metadata,
+	pub chunk_size: usize,
+	pub buffer_size: usize,
 }
 
 impl Push {
 	pub fn new(bucket_connection: BucketConnection, metadata: Metadata) -> Self {
-		Self { bucket_connection, metadata }
+		Self {
+			bucket_connection,
+			metadata,
+			chunk_size: DEFAULT_CHUNK_SIZE,
+			buffer_size: BUFFER_SIZE,
+		}
 	}
 
 	pub(crate) async fn upload_path(
@@ -31,6 +43,7 @@ impl Push {
 		let bucket = self.bucket_connection.bucket.clone();
 		let key =
 			format!("{}/{}", self.metadata.syncer_epoch_prefix()?, relative_path.to_string_lossy());
+		tracing::info!("Pushing file on S3 on bucket:{bucket} with key: {key}");
 		let body = ByteStream::from_path(full_path).await?;
 		let s3_path = format!("s3://{}/{}", bucket, key);
 		let output = self
@@ -87,7 +100,6 @@ impl Push {
 		// upload each file
 		let mut manifest_futures = Vec::new();
 		for (relative_path, full_path) in path_tuples {
-			tracing::info!("Pushing file:{:?}", relative_path);
 			let future = self.add_upload_entry(relative_path, full_path, None);
 			manifest_futures.push(future);
 		}
@@ -170,9 +182,24 @@ impl Push {
 #[async_trait::async_trait]
 impl PushOperations for Push {
 	async fn push(&self, package: Package) -> Result<Package, anyhow::Error> {
-		tracing::info!("Pushing package:{package:?}");
+		tracing::debug!("Pushing package:{package:?}");
 		// prune the old epochs
 		self.prune().await?;
+
+		// Split the too big files
+		let mut new_package_elements = vec![];
+		for element in package.0 {
+			for file_path in element.sync_files {
+				let new_files =
+					split_archive(file_path, &element.root_dir, self.chunk_size, self.buffer_size)?;
+				let mut new_element = PackageElement::new(element.root_dir.clone());
+				for dest in new_files {
+					new_element.add_sync_file(dest);
+				}
+				new_package_elements.push(new_element);
+			}
+		}
+		let package = Package(new_package_elements);
 
 		// upload the package
 		let mut manifest_futures = Vec::new();
@@ -183,4 +210,71 @@ impl PushOperations for Push {
 		let manifests = futures::future::try_join_all(manifest_futures).await?;
 		Ok(Package(manifests))
 	}
+}
+
+fn split_archive<P: AsRef<Path>>(
+	archive: PathBuf,
+	root_dir: P,
+	chunk_size: usize,
+	buffer_size: usize,
+) -> Result<Vec<PathBuf>, anyhow::Error> {
+	let output_dir = root_dir.as_ref();
+
+	// Check the file size before proceeding with the split
+	let file_metadata = std::fs::metadata(&archive)?;
+	let file_size = file_metadata.len() as usize;
+	if file_size <= chunk_size {
+		return Ok(vec![archive]);
+	}
+
+	let archive_file = File::open(&archive)?;
+
+	std::fs::create_dir_all(output_dir)?;
+
+	let mut chunk_num = 0;
+	let mut buffer = vec![0; buffer_size];
+
+	let archive_relative_path = archive.strip_prefix(&output_dir)?;
+	let mut input_reader = StdBufReader::new(archive_file);
+
+	let mut chunk_list = vec![];
+	loop {
+		// Create a new file for the chunk
+		let chunk_path = output_dir.join(format!(
+			"{}_{:03}.chunk",
+			archive_relative_path.to_string_lossy(),
+			chunk_num
+		));
+
+		let mut chunk_file = File::create(&chunk_path)?;
+
+		let mut all_read_bytes = 0;
+		let end = loop {
+			// Read a part of the chunk into the buffer
+			let bytes_read = input_reader.read(&mut buffer)?;
+			if bytes_read == 0 {
+				break true; // End of chunk file
+			}
+
+			// Write the buffer data to the output file
+			chunk_file.write_all(&buffer[..bytes_read])?;
+			all_read_bytes += bytes_read;
+			if all_read_bytes >= chunk_size {
+				break false;
+			}
+		};
+
+		if all_read_bytes == 0 {
+			break; // End of chunk file and discard the current one.
+		}
+
+		chunk_num += 1;
+		chunk_list.push(chunk_path);
+		if end {
+			break; // End of chunk file
+		}
+	}
+
+	tracing::info!("split_archive return {chunk_list:?}",);
+	Ok(chunk_list)
 }
